This is a small script to merge the root files on pnfs.
It takes as input a text file with the subdirectory paths on pnfs.
For each path, all the root files present in this directory are merged in a root file located on the m machines. Then it moves this file to the pnfs area.

BE SURE TO HAVE ENOUGH SPACE ON YOUR MX ACCOUNT !
Some trees are really big => you need to have 100 Gb of free space.


You need to replace lathomas by your username on the m machines.


To execute :
chmod u+x mergingonpnfs
./mergingonpnfs my_text_file_with_the_list_of_subdirectories_on_pnfs